Allow up to 32 byte bitstrings

+++ src/bacnet/bacstr.h
@@ -77,7 +77,7 @@ extern "C" {
     uint8_t bitstring_bytes_used(
         BACNET_BIT_STRING * bit_string);
     BACNET_STACK_EXPORT
-    uint8_t bitstring_bits_capacity(
+    unsigned bitstring_bits_capacity(
         BACNET_BIT_STRING * bit_string);
 /* used for encoding and decoding from the APDU */
     BACNET_STACK_EXPORT
